feat(driver/timer): hybrid timer, hw event scheduling
This change upgrades the timer API to include hardware event scheduling.

We also update our default nRF timer implementation to implement the new
API, notably to support hardware event scheduling.

The nRF timer was further improved as a hybrid sleep/high-resolution
timer.

The udp example is now using fixed scheduling instants as well to
benefit from improved timer precision and latency.

The udp-fast example was removed as it became redundant with the update
to the udp sample.

The remaining examples were upgraded to use the new timer API. We also
moved peripherals configuration into a common module as it is now much
more involved than before.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>