fix(driver): propagate driver state through fallback
So far we did not make use of inner driver state: The nRF driver used an
empty struct to represent its implementation.

This will change when we introduce the interrupt executor to the radio
driver.

In preparation we need to ensure that driver state will be propagated
through a fallback event, too.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>