feat(timer): make timer instantiable
The timer must be available in different places in the framework
including places where the implementation of the timer must be generic
over the driver configuration.

At the same time, the timer consumes resources and must therefore be
instantiated proving exclusive ownership of those resources.

To achieve both, we make the timer instantiable (in exchange for
resource ownership) and also Copy so that it can be shared between
framework components.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>